more updates with event renaming
