+++ packages/event-builder/src/__tests__/build-client-changed-event.spec.ts
