+++ packages/event-builder/src/event-builder.ts
@@ -2,7 +2,7 @@ import { randomUUID } from "node:crypto";
 import {
   $ClientChangedEvent,
   ClientChangedEvent,
-} from "@nhsdigital/nhs-notify-config-schemas/src/schemas/client-mutated-event";
+} from "@nhsdigital/nhs-notify-config-schemas/src/schemas/client-changed-event";
 import { ClientInput } from "./input";
 import { eventSource } from "./config";
 
@@ -17,7 +17,7 @@ export const buildEvent = (input: ClientInput): ClientChangedEvent => {
     subject: input.clientId,
     type: "uk.nhs.notify.config.ClientChanged.v1",
     dataschema:
-      "https://notify.nhs.uk/events/schemas/client-mutated/1.0.0.json",
+      "https://notify.nhs.uk/events/schemas/client-changed/1.0.0.json",
     dataschemaversion: "1.0.0",
     data: {
       id: input.clientId,
