+++ packages/schemas/src/__tests__/events/client-changed.test.ts
@@ -1,6 +1,6 @@
 import fs from 'node:fs';
 import path from 'node:path';
-import { $ClientChangedEvent } from '../../schemas/client-mutated-event';
+import { $ClientChangedEvent } from '../../schemas/client-changed-event';
 
 describe('ClientChangedEvent validations', () => {
