+++ packages/schemas/src/cli/generate-json.ts
@@ -1,12 +1,12 @@
 import { z } from 'zod';
 import * as fs from 'node:fs';
 import packageJson from '../../package.json';
-import { $ClientChangedEvent } from '../schemas/client-mutated-event';
+import { $ClientChangedEvent } from '../schemas/client-changed-event';
 
 const version = packageJson.version;
 
 for (const [key, schema] of Object.entries({
-  'client-mutated': $ClientChangedEvent
+  'client-changed': $ClientChangedEvent
 })) {
   const json = z.toJSONSchema(schema);
   const file = `json/${key}-${version}.json`;
