+++ packages/schemas/src/schemas/client-changed-event.ts
@@ -6,7 +6,7 @@ const $ClientChangedEventData = z.union([$ClientWithAPIM, $ClientWithMESH]);
 
 const $ClientChangedEventMetadata = $EventMetadata.extend({
   type: z.literal('uk.nhs.notify.config.ClientChanged.v1'),
-  dataschema: z.string().regex(/^https:\/\/notify\.nhs\.uk\/events\/schemas\/client-mutated\/1.\d+\.\d+\.json$/),
+  dataschema: z.string().regex(/^https:\/\/notify\.nhs\.uk\/events\/schemas\/client-changed\/1.\d+\.\d+\.json$/),
   dataschemaversion: z.string().regex(/^1\.\d+\.\d+$/), // Matches semantic versioning format with fixed major version
 });
